Refactor: Performance optimization and code structure improvements
Problem:
- Large JavaScript bundle size (234.81 kB)
- useFormatter hook was too complex (164 lines)
- Hard-coded sample data mixed with business logic
- Loose type definitions reducing type safety

Solution:
- Implemented dynamic imports for Prism.js with code splitting
- Split useFormatter into focused hooks (state, actions, info)
- Externalized sample data to constants file
- Improved type definitions with strict unions
- Added lazy loading for syntax highlighting components

Performance Improvements:
- Bundle size reduced from 234.81 kB to 214.16 kB (-20.65 kB)
- Gzipped size reduced from 74.00 kB to 67.23 kB (-6.77 kB)
- Prism.js now loads only when needed
- Code splitting for language-specific components

Code Quality Improvements:
- Split 164-line useFormatter into 4 focused hooks
- Better separation of concerns
- Stricter TypeScript type definitions
- Externalized constants for better maintainability
