+++ src/constants/sampleData.ts
@@ -1,0 +1,25 @@
+export const SAMPLE_JSON = `{
+  "name": "John Doe",
+  "age": 30,
+  "city": "Tokyo",
+  "hobbies": ["reading", "gaming"],
+  "address": {
+    "street": "123 Main St",
+    "zipCode": "100-0001"
+  }
+}`;
+
+export const SAMPLE_XML = `<?xml version="1.0" encoding="UTF-8"?>
+<person>
+  <name>John Doe</name>
+  <age>30</age>
+  <city>Tokyo</city>
+  <hobbies>
+    <hobby>reading</hobby>
+    <hobby>gaming</hobby>
+  </hobbies>
+  <address>
+    <street>123 Main St</street>
+    <zipCode>100-0001</zipCode>
+  </address>
+</person>`;
